# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SkyBooker API",
        "description": "Flight Booking System API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/v1/flights": {
            "get": {
                "tags": [
                    "Flights"
                ],
                "summary": "List all Flights with pagination",
                "operationId": "0181da5af73cf8fb4c93a2f04ce0fcf1",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of Flights with pagination"
                    }
                }
            },
            "post": {
                "tags": [
                    "Flights"
                ],
                "summary": "Create a new Flight",
                "operationId": "f6ebc8d1e7dfd1999e0354166fe9df1d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "flight_number",
                                    "airline_id",
                                    "origin_airport_id",
                                    "destination_airport_id",
                                    "airplane_id",
                                    "departure_date",
                                    "departure_time",
                                    "arrival_date",
                                    "arrival_time",
                                    "duration",
                                    "price",
                                    "total_seats",
                                    "status"
                                ],
                                "properties": {
                                    "flight_number": {
                                        "type": "string",
                                        "example": "AA123"
                                    },
                                    "airline_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "origin_airport_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "destination_airport_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "airplane_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "departure_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-08-20"
                                    },
                                    "departure_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "10:00:00"
                                    },
                                    "arrival_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-08-20"
                                    },
                                    "arrival_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "14:00:00"
                                    },
                                    "duration": {
                                        "type": "string",
                                        "example": "4h 30m"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 499.99
                                    },
                                    "total_seats": {
                                        "type": "integer",
                                        "example": 189
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "scheduled",
                                        "enum": [
                                            "scheduled",
                                            "open",
                                            "closing",
                                            "completed",
                                            "cancelled",
                                            "delayed",
                                            "boarding",
                                            "departed"
                                        ]
                                    },
                                    "booking_deadline": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-08-19",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Flight created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/flights/{flight}": {
            "get": {
                "tags": [
                    "Flights"
                ],
                "summary": "Get Flight by ID",
                "operationId": "346062cb1496c197566ba533d3f986db",
                "parameters": [
                    {
                        "name": "flight",
                        "in": "path",
                        "description": "Flight ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Flight retrieved successfully"
                    },
                    "404": {
                        "description": "Flight not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Flights"
                ],
                "summary": "Update a Flight",
                "operationId": "4fb080cc593b17ca9e4c7e1a76bfbe31",
                "parameters": [
                    {
                        "name": "flight",
                        "in": "path",
                        "description": "Flight ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "flight_number": {
                                        "type": "string",
                                        "example": "AA456"
                                    },
                                    "airline_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "origin_airport_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "destination_airport_id": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "airplane_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "departure_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-08-21"
                                    },
                                    "departure_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "11:00:00"
                                    },
                                    "arrival_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-08-21"
                                    },
                                    "arrival_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "15:00:00"
                                    },
                                    "duration": {
                                        "type": "string",
                                        "example": "4h 00m"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 599.99
                                    },
                                    "total_seats": {
                                        "type": "integer",
                                        "example": 200
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "boarding",
                                        "enum": [
                                            "scheduled",
                                            "open",
                                            "closing",
                                            "completed",
                                            "cancelled",
                                            "delayed",
                                            "boarding",
                                            "departed"
                                        ]
                                    },
                                    "booking_deadline": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-08-20",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Flight updated successfully"
                    },
                    "404": {
                        "description": "Flight not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Flights"
                ],
                "summary": "Delete a Flight",
                "operationId": "d1de382152a384d9f0735ea83552b3e5",
                "parameters": [
                    {
                        "name": "flight",
                        "in": "path",
                        "description": "Flight ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Flight deleted successfully"
                    },
                    "404": {
                        "description": "Flight not found"
                    }
                }
            }
        },
        "/api/v1/flights/{flight}/bookings": {
            "get": {
                "tags": [
                    "Flights"
                ],
                "summary": "Get all bookings for a flight",
                "operationId": "d3d9495ffc389cb86e1450f5431168df",
                "parameters": [
                    {
                        "name": "flight",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of bookings for the flight"
                    },
                    "404": {
                        "description": "Flight not found"
                    }
                }
            }
        },
        "/api/v1/flights/{flight}/tickets": {
            "get": {
                "tags": [
                    "Flights"
                ],
                "summary": "Get all tickets for a flight",
                "operationId": "0022677820e760e030cabf10780ff1e9",
                "parameters": [
                    {
                        "name": "flight",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tickets for the flight"
                    },
                    "404": {
                        "description": "Flight not found"
                    }
                }
            }
        },
...
    },
...
}